requirements.txt
redis
python-dotenv
fastapi
pydantic